The row-lock handlers' re-entry rule is pinned, and the page allows the counted variant
`how-tos/lock-handler.md` calls re-entry "the single most important rule", and
nothing asserted it for either shipped row-lock family: the six unit files that
acquire a lock never acquire twice from one requestor against one. The rule
lives in `DbLockHandler`, so one case over the fixture's existing source covers
`UpdateRowLockHandler`, `SelectForUpdateLockHandler` and the two dialect
subclasses that inherit it.

The statement count comes from a subclass that counts `ExecuteSql` rather than
from a counting provider: the store mints commands from the concrete provider
type, so a decorator would fall into the fallback branch and be handed a
command attached to no connection. Beside it, the two halves that say what
`false` is about — the two lock kinds are held independently, another requestor
is told `true`, and a stranger's release neither throws nor unseats the holder.

The page stated the rule more absolutely than the shipped handlers keep it:
`ILockHandler` allows a handler that counts its holds to answer `true`, and
`SqliteLockHandler` does. The page now says so and names it, so a reader
writing a counted handler is not told they are breaking the rule.

The fixture is renamed for what it now covers; cancellation was only ever half
of it.
